requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
transformers
torch
faiss-cpu
pyyaml
